* glib/Value.cs: add new constructors for enum and boxed values         that take the name of the type rather than an object/property         name pair; this way they work for both GObject properties and         GtkContainer child properties.
        * glib/glue/value.c (gtksharp_value_create_from_type_name): glue
        for that

        * glib/Opaque.cs (GetOpaque): Fix this.

        * generator/Property.cs (Generate): Use the new GLib.Value
        constructors. (Fixes setting of enum-valued child properties.)

svn path=/trunk/gtk-sharp/; revision=36174

+++ glib/Opaque.cs
@@ -54,7 +54,7 @@ namespace GLib {
 
 		public static Opaque GetOpaque(IntPtr o)
 		{
-			Opaque obj = (Opaque)Opaques[(int)o];
+			Opaque obj = (Opaque)Opaques[o];
 			if (obj != null) return obj;
 			return null; //FIXME: Call TypeParser here eventually.
 		}
